Split external wallet actions into outgoing and incoming
