+++ src/wallet/interface/outgoing.ts
@@ -1,0 +1,141 @@
+// FUNDING
+// =======
+
+export const FUNDING_SUCCESS = 'WALLET.FUNDING.SUCCESS';
+export const FUNDING_FAILURE = 'WALLET.FUNDING.FAILURE';
+
+export const fundingSuccess = channelId => ({
+  type: FUNDING_SUCCESS as typeof FUNDING_SUCCESS,
+  channelId,
+});
+export const fundingFailure = (channelId, reason) => ({
+  type: FUNDING_FAILURE as typeof FUNDING_FAILURE,
+  channelId,
+  reason,
+});
+
+export type FundingSuccess = ReturnType<typeof fundingSuccess>;
+export type FundingFailure = ReturnType<typeof fundingFailure>;
+export type FundingResponse = FundingSuccess | FundingFailure;
+
+// CHANNELS
+// ========
+
+export const CHANNEL_OPENED = 'WALLET.CHANNEL.OPENED';
+export const CHANNEL_CLOSED = 'WALLET.CHANNEL.CLOSED';
+
+export const channelOpened = (channelId: string) => ({
+  type: CHANNEL_OPENED as typeof CHANNEL_OPENED,
+  channelId,
+});
+export const channelClosed = (walletId: string) => ({
+  type: CHANNEL_CLOSED as typeof CHANNEL_CLOSED,
+  walletId,
+});
+
+export type ChannelOpened = ReturnType<typeof channelOpened>;
+export type ChannelClosed = ReturnType<typeof channelClosed>;
+
+// VALIDATION
+// ==========
+
+export const VALIDATION_SUCCESS = 'WALLET.VALIDATION.SUCCESS';
+export const VALIDATION_FAILURE = 'WALLET.VALIDATION.FAILURE';
+
+export const validationSuccess = (requestId: string) => ({
+  type: VALIDATION_SUCCESS as typeof VALIDATION_SUCCESS,
+  requestId,
+});
+export const validationFailure = (requestId: string, reason: string) => ({
+  type: VALIDATION_FAILURE as typeof VALIDATION_FAILURE,
+  requestId,
+  reason,
+});
+
+export type ValidationSuccess = ReturnType<typeof validationSuccess>;
+export type ValidationFailure = ReturnType<typeof validationFailure>;
+export type ValidationResponse = ValidationSuccess | ValidationFailure;
+
+
+// SIGNATURE
+// =========
+
+export const SIGNATURE_SUCCESS = 'WALLET.SIGNATURE.SUCCESS';
+export const SIGNATURE_FAILURE = 'WALLET.SIGNATURE.FAILURE';
+
+export const signatureSuccess = (requestId: string, signature: string) => ({
+  type: SIGNATURE_SUCCESS as typeof SIGNATURE_SUCCESS,
+  requestId,
+  signature,
+});
+export const signatureFailure = (requestId: string, reason: string) => ({
+  type: SIGNATURE_FAILURE as typeof SIGNATURE_FAILURE,
+  requestId,
+  reason,
+});
+
+export type SignatureSuccess = ReturnType<typeof signatureSuccess>;
+export type SignatureFailure = ReturnType<typeof signatureFailure>;
+export type SignatureResponse = SignatureSuccess | SignatureFailure;
+
+
+// WITHDRAWAL
+// ==========
+
+export const WITHDRAWAL_SUCCESS = 'WALLET.WITHDRAWAL.SUCCESS';
+export const WITHDRAWAL_FAILURE = 'WALLET.WITHDRAWAL.FAILURE';
+
+export const withdrawalSuccess = transaction => ({
+  type: WITHDRAWAL_SUCCESS as typeof WITHDRAWAL_SUCCESS,
+  transaction,
+});
+export const withdrawalFailure = (reason) => ({
+  type: WITHDRAWAL_FAILURE as typeof WITHDRAWAL_FAILURE,
+  reason,
+});
+
+export type WithdrawalSuccess = ReturnType<typeof withdrawalSuccess>;
+export type WithdrawalFailure = ReturnType<typeof withdrawalFailure>;
+export type WithdrawalResponse = WithdrawalSuccess | WithdrawalFailure;
+
+
+// INITIALIZATION
+// ==============
+
+export const INITIALIZATION_SUCCESS = 'WALLET.INITIALIZATION.SUCCESS';
+export const INITIALIZATION_FAILURE = 'WALLET.INITIALIZATION.FAILURE';
+
+export const initializationSuccess = address => ({
+  type: INITIALIZATION_SUCCESS as typeof INITIALIZATION_SUCCESS,
+  address,
+});
+
+export const initializationFailure = (message: string) => ({
+  type: INITIALIZATION_FAILURE as typeof INITIALIZATION_FAILURE,
+  message,
+});
+
+export type InitializationSuccess = ReturnType<typeof initializationSuccess>;
+
+
+// MESSAGING
+// =========
+export const SEND_MESSAGE = 'WALLET.MESSAGING.SEND';
+
+export const sendMessage = (to: string, data: string) => ({
+  type: SEND_MESSAGE as typeof SEND_MESSAGE,
+  to,
+  data,
+});
+
+export type SendMessage = ReturnType<typeof sendMessage>;
+
+export const MESSAGE_RECEIVED = 'WALLET.MESSAGING.MESSAGE_RECEIVED';
+export const messageReceived = (positionData: string, signature: string) => ({
+  type: MESSAGE_RECEIVED as typeof MESSAGE_RECEIVED,
+  positionData,
+  signature,
+});
+
+
+export type MessageReceived = ReturnType<typeof messageReceived>;
